requirements.txt
asgiref==3.6.0
Django==4.0
django-ckeditor==6.5.1
django-filter==23.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-phonenumber-field==7.0.2
djangorestframework==3.13.0
djangorestframework-simplejwt==5.2.2
phonenumbers==8.13.10
Pillow==9.5.0
PyJWT==2.6.0
pytz==2023.3
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
tzdata==2023.3
